% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{addFeatureMatrix}
\alias{addFeatureMatrix}
\title{Add feature matrix}
\usage{
addFeatureMatrix(feature_mv, data_df, option_num, feature_name, id = NULL)
}
\arguments{
\item{feature_mv}{The features to be added}

\item{data_df}{The raw data.frame}

\item{option_num}{Number of options, should be the same as the column number of the  feaature matrix}

\item{feature_name}{Name of feature to appear in formatted data}

\item{id}{Indicating which rows to extract from the feature matrix for each row of the raw data frame}
}
\description{
Add feature matrix
}
