% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{changeBetaNames}
\alias{changeBetaNames}
\title{Change variable names}
\usage{
changeBetaNames(names, beta, Q, S)
}
\arguments{
\item{names}{A vector of characters. Variables taken from rstan outputs.}

\item{beta}{Variable names to be included in the memory model, e.g., c('sameC','sameA','dist').
Note that in that case, columns 'sameC_1'...'sameC_C' and 'sameA_1',...,'sameA_C'  and 'dist_1',...,'dist_C' should be specified.
In the memory model, we could include sameC (cluster resampling; sameC_value== 1), sameA (decision congruence; sameA_value== 1),
and dist (semantic congruence; dist_value == 1)}

\item{Q}{Number of questions}

\item{S}{Number of participants}
}
\description{
Change variable names
}
