% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{getData}
\alias{getData}
\title{Format data for the thoughtListing project}
\usage{
getData(cluster_num = NULL, data = NULL, dist = NULL)
}
\arguments{
\item{cluster_num}{number of thought cluster for one option. If 3 then option number is 2 * 3 + 1 = 7.}

\item{data}{thought listing data read from .csv}

\item{dist}{semantic distance data read from .csv}
}
\description{
Add feature matrix
(decision congruence, semantic congruence and cluster resmapling matrix)
to the thought listing data.
}
