% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{getFormatData}
\alias{getFormatData}
\title{Format data for the thoughtListing project}
\usage{
getFormatData(data_all, dist_all, cluster_num = NULL)
}
\arguments{
\item{data_all}{dataframe witth qID, sID, tNo, and cID}

\item{dist_all}{distance matrix}

\item{cluster_num}{number of thought cluster for one option. If 3 then option number is 2 * 3 + 1 = 7.}
}
\description{
Add feature matrix
(decision congruence, semantic congruence and cluster resmapling matrix)
to the thought listing data.
}
