% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{getModelDiag}
\alias{getModelDiag}
\title{Model diagnostics.}
\usage{
getModelDiag(stan_data_fit, csv_name = NULL, dev_name = "dev_m")
}
\arguments{
\item{stan_data_fit}{model outputs obtained from getStanFit}

\item{csv_name}{if specified then outputs will be saved in a .csv file starting with this name}

\item{dev_name}{string. name for dev in rstan codes}
}
\value{
A dataframe indicating DIC, WAIC, model running time,
maximum rhat, minimum effective sample size,
number of divergent transitions and number of max treedepth hit.
}
\description{
\code{getModelDiag} returns model diagnostics,
and save the results to a .csv file when \code{csv_name} is specified
}
