% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{getPara}
\alias{getPara}
\title{Parameter summaries}
\usage{
getPara(
  para_name,
  stan_data_fit,
  csv_name = NULL,
  quantiles = c(0.025, 0.05, 0.1, 0.2, 0.25, 0.5, 0.75, 0.8, 0.9, 0.95, 0.975)
)
}
\arguments{
\item{para_name}{parameter to extract}

\item{stan_data_fit}{model outputs obtained from getStanFit}

\item{csv_name}{if specified then outputs will be saved in a .csv file starting with this name}

\item{quantiles}{quantiles to extract}
}
\value{
A dataframe with parameter mean, quantiles and sd, can be saved to a csv file
}
\description{
\code{getPara} returns parameter mean, quantiles and sd
and save the results to a .csv file when \code{csv_name} is specified
}
