% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_function.R
\name{getSimulation}
\alias{getSimulation}
\title{Model recovery}
\usage{
getSimulation(
  cluster_num = 3,
  sim_total = NULL,
  X_sim = NULL,
  deltaM_value = 8,
  deltaD_value = 8,
  ar_value = NULL,
  max_tNo_prd = 30,
  seed_no = seed_no,
  deltaM_true = NULL,
  alpha_true = NULL,
  beta_true = NULL,
  deltaD_true = NULL,
  threshold_true = NULL,
  sigma_true = NULL,
  iter_num = NULL,
  chain_num = NULL,
  warmup_num = NULL,
  core_num = NULL,
  refresh = 500
)
}
\arguments{
\item{X_sim}{Feature matrix, for different clusters. Needed for simulation.}

\item{deltaM_value}{1 minus memory decay; 0 for full decay (marcov property); 1 for no decay;
8 for estimating it as a free parameter}

\item{deltaD_value}{1 minus decision decay; 0 for full decay (marcov property); 1 for no decay;
8 for estimating it as a free parameter}

\item{ar_value}{0 for absolute accumulation; 1 for relative accumulation}

\item{max_tNo_prd}{Maximum length of a simulated trial}

\item{seed_no}{Seed number}

\item{deltaM_true}{1 minus decay for memory}

\item{alpha_true}{baseline activations for memory}

\item{beta_true}{feature coefficients}

\item{deltaD_true}{1 minus decay for decision}

\item{threshold_true}{threshold for decision}

\item{sigma_true}{noise for decision}

\item{iter_num}{Iteration number}

\item{chain_num}{Chain number}

\item{warmup_num}{Warm-up or burn-in number}

\item{core_num}{Number of cores to be used}

\item{refresh}{Default 500}
}
\description{
\code{getSimulation} simulates data and fit memory and decision model to the simulated dataset.
}
