% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{getStanData}
\alias{getStanData}
\title{Format data for Stan}
\usage{
getStanData(
  beta = character(0),
  deltaM_value = NULL,
  option_num = 3 * 2 + 1,
  format_data
)
}
\arguments{
\item{beta}{variable names to be included in the model, e.g., c('beta','gamma').
note that in the formatted data, columns 'beta_1'...'beta_C' and 'gamma_1',...,'gamma_C' should be specified}

\item{deltaM_value}{0 for full decay (marcov property); 1 for no decay;
9 for estimating it as a free parameter}

\item{option_num}{number of total options in each choice (C)}

\item{format_data}{this should be a formated data.frame. sID: participant IDs. qID: question ID.
cID: option ID being sampled (can take 1 to C). tNo: thought number in that question.
terminate: 0 if continue sampling (not necessary)}
}
\description{
Format data for Stan
}
