% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{getStanFit}
\alias{getStanFit}
\title{Fit the model!}
\usage{
getStanFit(
  beta = NULL,
  deltaM_value = NULL,
  binary_value = NULL,
  ar_value = NULL,
  random_value = NULL,
  deltaD_value = NULL,
  option_num,
  format_data,
  save_model_file = NULL,
  init_values = "random",
  iter_num,
  chain_num,
  warmup_num,
  core_num,
  adapt_delta = 0.999,
  stepsize = 0.01,
  max_treedepth = 20,
  save_warmup = FALSE,
  refresh = 500,
  init_r = 0.5,
  hier_value = 1,
  X_sim = NULL,
  cluster_rating_m = NULL,
  max_tNo_prd = NULL
)
}
\arguments{
\item{beta}{Variable names to be included in the memory model, e.g., c('sameC','sameA','dist').
Note that in that case, columns 'sameC_1'...'sameC_C' and 'sameA_1',...,'sameA_C'  and 'dist_1',...,'dist_C' should be specified.
In the memory model, we could include sameC (cluster resampling; sameC_value== 1), sameA (decision congruence; sameA_value== 1),
and dist (semantic congruence; dist_value == 1)}

\item{deltaM_value}{1-memory_decay; 0 for full decay (marcov property); 1 for no decay;
8 for estimating it as a free parameter}

\item{binary_value}{0 for continuous; 1 for discrete inputs}

\item{ar_value, }{0 for absolute accumulation; 1 for relative accumulation}

\item{random_value, }{0 for time limit; 1 for threhold-based}

\item{deltaD_value, }{1-decision_decay; 0 for full decay (marcov property); 1 for no decay;
8 for estimating it as a free parameter}

\item{option_num}{Number of total options in each choice (C)}

\item{format_data}{This should be a formated data.frame. sID: participant IDs. qID: question ID.
cID: option ID being sampled (can take 1 to C). tNo: thought number in that question.
terminate: 0 if continue sampling (not necessary)}

\item{save_model_file}{Specify a name if you want to save the RStan samples and diagnostics to a csv file}

\item{init_values}{Default is 'random'; can be specified using a function or a list}

\item{iter_num}{Iteration number}

\item{chain_num}{Chain number}

\item{warmup_num}{Warm-up or burn-in number}

\item{core_num}{Number of cores to be used. equal or less to chain_num}

\item{adapt_delta}{Default 0.999}

\item{stepsize}{Default 0.01}

\item{max_treedepth}{Default 20}

\item{refresh}{Default 500}

\item{hier_value}{Whether to use a hiearchical model. Default is 1.
Note that the hierachical version only work for decision or memory models seperately.}
}
\value{
Use $stan_fit to access the returned object from Stan;
Use $stan_data to see the input data of the Stan model
}
\description{
\code{getStanFit} use RStan to fit a memory model or a decision model.
A memory model will be fit if deltaM_value is specified;
A decision model will be fit if deltaD_value is specified;
If both specified then the two models will be fit simultaneously, and PPC will be run
Note that for memory models the option number (C) should be the same across all questions.
}
