% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_function.R
\name{simulateChoice}
\alias{simulateChoice}
\title{Simulate option}
\usage{
simulateChoice(
  beta_true,
  delta_true,
  alpha_true,
  feature_data,
  option_num,
  question_num,
  seed_no = 1
)
}
\arguments{
\item{beta_true}{Values of beta vector (feature * 1)}

\item{delta_true}{Values of delta}

\item{alpha_true}{Values of alpha}

\item{feature_data}{feature matrix (option * feature);
dependent on the last chosen option}

\item{option_num}{Number of total options in each choice (C)}

\item{question_num}{totaal simulation number}
}
\description{
Feature mattrix is dependent on the option sampled at the last time point
}
