% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{addFeatureMatrix}
\alias{addFeatureMatrix}
\title{Title}
\usage{
addFeatureMatrix(feature_mv, data_df, option_num, feature_name, id = NULL)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Title
}
