% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{changeBetaNames}
\alias{changeBetaNames}
\title{Change variable names}
\usage{
changeBetaNames(names, beta, Q, S)
}
\arguments{
\item{names}{A vector of characters. Variables taken from rstan outputs.}

\item{beta}{Variable names to be included in the model, e.g., c('beta','gamma').
Note that in the formatted data, columns 'beta_1'...'beta_C' and 'gamma_1',...,'gamma_C' should be specified}

\item{Q}{Number of questions}

\item{S}{Number of participants}
}
\description{
Change variable names
}
