% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{getFeatureMatrices}
\alias{getFeatureMatrices}
\title{Title}
\usage{
getFeatureMatrices(
  beta,
  format_data,
  deltaM_value,
  option_num,
  print_names = F
)
}
\arguments{
\item{beta}{Variable names to be included in the model, e.g., c('beta','gamma').
Note that in the formatted data, columns 'beta_1'...'beta_C' and 'gamma_1',...,'gamma_C' should be specified}

\item{format_data}{This should be a formated data.frame. sID: participant IDs. qID: question ID.
cID: option ID being sampled (can take 1 to C). tNo: thought number in that question.
terminate: 0 if continue sampling (not necessary)}

\item{deltaM_value}{0 for full decay (marcov property); 1 for no decay;
9 for estimating it as a free parameter}

\item{option_num}{Number of total options in each choice (C)}

\item{print_names}{whether to include column and row namaes; default FALSE}
}
\description{
Title
}
