% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_function.R
\name{getSimulation}
\alias{getSimulation}
\title{Model recovery}
\usage{
getSimulation(
  cluster_num = 3,
  sim_total = NULL,
  X_sim = NULL,
  deltaM_value = 8,
  deltaD_value = 8,
  ar_value = NULL,
  max_tNo_prd = 30,
  seed_no = seed_no,
  deltaM_true = NULL,
  alpha_true = NULL,
  beta_true = NULL,
  deltaD_true = NULL,
  threshold_true = NULL,
  sigma_true = NULL,
  iter_num = NULL,
  chain_num = NULL,
  warmup_num = NULL,
  core_num = NULL,
  refresh = 500
)
}
\arguments{
\item{deltaM_value}{1-memory_decay; 0 for full decay (marcov property); 1 for no decay;
8 for estimating it as a free parameter}

\item{deltaD_value}{1-decision_decay; 0 for full decay (marcov property); 1 for no decay;
8 for estimating it as a free parameter}

\item{ar_value}{0 for absolute accumulation; 1 for relative accumulation}

\item{deltaM_true}{1-decay for memory}

\item{alpha_true}{baserate for memory}

\item{beta_true}{length 3; for cluster resampling, decision congruence and semantic congruence respectively}

\item{deltaD_true}{1-decay for decision}

\item{threshold_true}{threshold for decision}

\item{sigma_true}{noise for decision}

\item{iter_num}{Iteration number}

\item{chain_num}{Chain number}

\item{warmup_num}{Warm-up or burn-in number}

\item{core_num}{Number of cores to be used. equal or less to chain_num}

\item{refresh}{Default 500}
}
\description{
Simulate data and fit memory and decision model
}
