% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{getStanCode}
\alias{getStanCode}
\title{Retrieve stan code}
\usage{
getStanCode(hier_value = 1)
}
\arguments{
\item{hier_value}{Whether use hiearchical model. Default is 1.}
}
\description{
Retrieve stan code
}
