% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{getStanCode}
\alias{getStanCode}
\title{Retrieve stan code}
\usage{
getStanCode(
  deltaM_value = NULL,
  deltaD_value = NULL,
  hier_value = NULL,
  random_value = NULL
)
}
\arguments{
\item{deltaM_value}{1-memory_decay; 0 for full decay (marcov property); 1 for no decay;
8 for estimating it as a free parameter}

\item{deltaD_value}{1-decision_decay; 0 for full decay (marcov property); 1 for no decay;
8 for estimating it as a free parameter}

\item{hier_value}{Whether to use a hiearchical model. Default is 1.
Note that the hierachical version only work for decision or memory models seperately.}

\item{random_value}{0 for time limit; 1 for threhold-based}
}
\description{
Retrieve stan code
}
