% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{getStanData}
\alias{getStanData}
\title{Format data for Stan}
\usage{
getStanData(
  beta = character(0),
  deltaM_value = NULL,
  binary_value = NULL,
  ar_value = NULL,
  random_value = NULL,
  deltaD_value = NULL,
  option_num = 3 * 2 + 1,
  format_data,
  X_sim = NULL,
  cluster_rating_m = NULL,
  max_tNo_prd = NULL
)
}
\arguments{
\item{beta}{Variable names to be included in the memory model, e.g., c('sameC','sameA','dist').
Note that you have to include the associated columns in the data.
For example, to fit sameC, you have to include columns with names of 'sameC_1'...'sameC_C'.
To fit sameA, you have to include 'sameA_1',...,'sameA_C'.
In the memory model, we could include sameC (cluster resampling; sameC_value== 1), sameA (decision congruence; sameA_value== 1),
and dist (semantic congruence; dist_value == 1)}

\item{deltaM_value}{1 minus memory decay; 0 for full decay (marcov property); 1 for no decay;
8 for estimating it as a free parameter}

\item{binary_value}{0 for continuous inputs; 1 for discrete inputs}

\item{ar_value}{0 for absolute accumulation; 1 for relative accumulation}

\item{random_value}{0 for threhold-based;  1 for time limit;}

\item{deltaD_value}{1 minus decision decay; 0 for full decay (marcov property); 1 for no decay;
8 for estimating it as a free parameter}

\item{option_num}{Number of total clusters in each choice (C)}

\item{format_data}{This should be a formated data.frame.
sID: participant IDs.
qID: question ID.
cID: option ID being sampled (can take 1 to C).
tNo: thought number in that trial (from 1 to T).
terminate: 0 if continue sampling; 1 if yes; 0 if no (only needed for decision models)}

\item{X_sim}{Feature matrix, for different clusters. Needed for simulation.}

\item{cluster_rating_m}{Base rates of different ratings (-3 to 3) for each cluster. Needed for simulation.}

\item{max_tNo_prd}{Maximum length of a simulated trial}
}
\description{
Format data for Stan
}
