% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{getStanFit}
\alias{getStanFit}
\title{Fit the model!}
\usage{
getStanFit(
  beta,
  deltaM_value,
  option_num,
  format_data,
  save_model_file,
  init_values = "random",
  iter_num,
  chain_num,
  warmup_num,
  core_num,
  adapt_delta = 0.999,
  stepsize = 0.01,
  max_treedepth = 20,
  save_warmup = FALSE,
  refresh = 500,
  init_r = 0.5
)
}
\arguments{
\item{beta}{Variable names to be included in the model, e.g., c('beta','gamma').
Note that in the formatted data, columns 'beta_1'...'beta_C' and 'gamma_1',...,'gamma_C' should be specified}

\item{deltaM_value}{0 for full decay (marcov property); 1 for no decay;
9 for estimating it as a free parameter}

\item{option_num}{Number of total options in each choice (C)}

\item{format_data}{This should be a formated data.frame. sID: participant IDs. qID: question ID.
cID: option ID being sampled (can take 1 to C). tNo: thought number in that question.
terminate: 0 if continue sampling (not necessary)}

\item{save_model_file}{Specify a name if you want to save the RStan samples and diagnostics to a csv file}

\item{init_values}{Default is 'random'; can be specified using a function or a list}

\item{iter_num}{Iteration number}

\item{chain_num}{Chain number}

\item{warmup_num}{Warm-up or burn-in number}

\item{core_num}{Number of cores to be used. equal or less to chain_num}

\item{adapt_delta}{Default 0.999}

\item{stepsize}{Default 0.01}

\item{max_treedepth}{Default 20}

\item{refresh}{Default 500}
}
\value{
Use $stan_fit to access the returned object from Stan;
Use $stan_data to see the input data of the Stan model
}
\description{
\code{getStanFit} use RStan to fit your information sampling model.
Note that the option number (C) should be the same across all questions.
}
