% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{getStanFit}
\alias{getStanFit}
\title{Fit the model!}
\usage{
getStanFit(
  beta = NULL,
  deltaM_value = NULL,
  binary_value = NULL,
  ar_value = NULL,
  random_value = NULL,
  deltaD_value = NULL,
  option_num,
  format_data,
  save_model_file = NULL,
  init_values = "random",
  iter_num,
  chain_num,
  warmup_num,
  core_num,
  adapt_delta = 0.999,
  stepsize = 0.01,
  max_treedepth = 20,
  seed_no = 1,
  save_warmup = FALSE,
  refresh = 500,
  init_r = 0.5,
  hier_value = 1,
  X_sim = NULL,
  cluster_rating_m = NULL,
  max_tNo_prd = NULL
)
}
\arguments{
\item{beta}{Variable names to be included in the memory model, e.g., c('sameC','sameA','dist').
Note that you have to include the associated columns in the data.
For example, to fit sameC, you have to include columns with names of 'sameC_1'...'sameC_C'.
To fit sameA, you have to include 'sameA_1',...,'sameA_C'.
In the memory model, we could include sameC (cluster resampling; sameC_value== 1), sameA (decision congruence; sameA_value== 1),
and dist (semantic congruence; dist_value == 1)}

\item{deltaM_value}{1 minus memory decay; 0 for full decay (marcov property); 1 for no decay;
8 for estimating it as a free parameter}

\item{binary_value}{0 for continuous inputs; 1 for discrete inputs}

\item{ar_value, }{0 for absolute accumulation; 1 for relative accumulation}

\item{random_value, }{0 for threhold-based;  1 for time limit;}

\item{deltaD_value, }{1 minus decision decay; 0 for full decay (marcov property); 1 for no decay;
8 for estimating it as a free parameter}

\item{option_num}{Number of total clusters in each choice (C)}

\item{format_data}{This should be a formated data.frame.
sID: participant IDs.
qID: question ID.
cID: option ID being sampled (can take 1 to C).
tNo: thought number in that trial (from 1 to T).
terminate: 0 if continue sampling; 1 if yes; 0 if no (only needed for decision models)}

\item{save_model_file}{Specify a name if you want to save the RStan samples and diagnostics to a csv file.
This is sample_file for rstan::sampling}

\item{init_values}{Default is 'random'; can be specified using a function or a list}

\item{iter_num}{Iteration number}

\item{chain_num}{Chain number}

\item{warmup_num}{Warm-up or burn-in number}

\item{core_num}{Number of cores to be used}

\item{adapt_delta}{Default 0.999}

\item{stepsize}{Default 0.01}

\item{max_treedepth}{Default 20}

\item{seed_no}{Seed number}

\item{refresh}{Default 500}

\item{hier_value}{0 for fitting a non-hierarchical model; 1 for fitting a hierarchical model. Default is 1.
Note that the hierachical version only works for decision or memory models seperately.}

\item{X_sim}{Feature matrix, for different clusters. Needed for simulation.}

\item{cluster_rating_m}{Base rates of different ratings (-3 to 3) for each cluster. Needed for simulation.}

\item{max_tNo_prd}{Maximum length of a simulated trial}
}
\value{
Use $stan_fit to access the returned object from Stan;
Use $stan_data to see the input data of the Stan model
}
\description{
\code{getStanFit} uses RStan to fit a memory model or a decision model.
A memory model will be fit if deltaM_value is specified.
A decision model will be fit if deltaD_value is specified.
If both specified then the two models will be fit simultaneously, and posterior predictions will be simulated
Note that for memory models the option number (C) should be the same across all questions.
}
